requirements.txt
absl-py
accelerate
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
appnope
asttokens
astunparse
attrs
beautifulsoup4
blinker
cachetools
certifi
charset-normalizer
click
cmake
comm
contourpy
cycler
dash
dash-bootstrap-components
datasets
debugpy
decorator
dill
dlib
docx2txt
exceptiongroup
executing
face-recognition
face_recognition_models
facenet-pytorch
fastapi
ffmpy
filelock
fire
Flask
flatbuffers
fonttools
frozenlist
fsspec
gast
gdown
gitdb
GitPython
google-pasta
gradio
gradio_client
groovy
grpcio
gunicorn
h11
h5py
httpcore
httpx
huggingface
huggingface-hub
idna
imageio
imageio-ffmpeg
importlib_metadata
ipykernel
ipython
ipython_pygments_lexers
ipywidgets
itsdangerous
jedi
Jinja2
joblib
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
jupyterlab_widgets
kagglehub
keras
kiwisolver
lazy_loader
libclang
lightgbm
lime
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
ml_dtypes
mpmath
multidict
multiprocess
namex
narwhals
nest_asyncio
networkx
nlpaug
nltk
numpy
opencv-python
opt_einsum
optree
orjson
packaging
pandas
parso
pexpect
pickleshare
pillow
platformdirs
plotly
praw
prawcore
prompt_toolkit
propcache
protobuf
psutil
ptyprocess
pure_eval
pyarrow
pydantic
pydantic_core
pydeck
pydub
Pygments
pyparsing
PyPDF2
PySocks
python-dateutil
python-dotenv
python-env
python-multipart
pytz
PyYAML
pyzmq
referencing
regex
requests
retrying
rich
rpds-py
ruff
safehttpx
safetensors
scikit-image
scikit-learn
scipy
seaborn
semantic-version
sentencepiece
shellingham
six
smmap
sniffio
soupsieve
SQLAlchemy
sqlmodel
stack_data
starlette
stopwordsiso
streamlit
stringcase
sympy
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
tensorflow-text
termcolor
textblob
threadpoolctl
tifffile
tokenizers
toml
tomlkit
torch
torchaudio
torchvision
tornado
tqdm
traitlets
transformers
typer
typing_extensions
tzdata
update-checker
urllib3
uvicorn
watchdog
wcwidth
websocket-client
websockets
Werkzeug
widgetsnbextension
wordcloud
wrapt
xgboost
xxhash
yarl
zipp
